+++ node/libs/protobuf_build/src/lib.rs
@@ -142,18 +142,16 @@ pub struct Config {
 impl Config {
     /// Generates implementation of `prost_reflect::ReflectMessage` for a rust type generated
     /// from a message of the given `proto_name`.
-    fn reflect_impl(&self, proto_name: &ProtoName) -> anyhow::Result<Vec<syn::Item>> {
+    fn reflect_impl(&self, proto_name: &ProtoName) -> anyhow::Result<syn::Item> {
         let rust_name = proto_name
             .relative_to(&self.proto_root.to_name().context("invalid proto_root")?)
             .unwrap()
             .to_rust_type()?;
         let proto_name = proto_name.to_string();
         let protobuf_crate = &self.protobuf_crate;
-        Ok(vec![
+        Ok(
             syn::parse_quote! { #protobuf_crate::build::impl_reflect_message!(#rust_name, &DESCRIPTOR, #proto_name); },
-            syn::parse_quote! { #protobuf_crate::build::impl_serde_serialize!(#rust_name); },
-            syn::parse_quote! { #protobuf_crate::build::impl_serde_deserialize!(#rust_name); },
-        ])
+        )
     }
 
     /// Validates this configuration.
@@ -361,12 +359,10 @@ impl Config {
 
         // Generate the reflection code.
         for proto_name in extract_message_names(&descriptor) {
-            let impls = self
+            let item = self
                 .reflect_impl(&proto_name)
                 .with_context(|| format!("reflect_impl({proto_name})"))?;
-            for i in impls {
-                output.append_item(i);
-            }
+            output.append_item(item);
         }
 
         // Save output.
